refactor and close temp file
